TEST workaround to get older dxsync module working (partially w/out precompiled pathwatcher) on recent Node versions.
This is a potential test workaround, to get the older dxsync module working partially (without the precompiled pathwatcher support) on more recent versions of Node.   This is basically a copy of the 1.0.2 patch release,  with pathwatcher removed and with the gracefulfs dependency updated from version 3.x to version 4.x.    Not fully tested yet, provided as-is, for those who can't get dxsync to work at all on Node 8.x, a partial workaround to try out.